requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Environment and Configuration
python-dotenv==1.0.0

# YouTube and Media Processing
yt-dlp==2023.11.16
ffmpeg-python==0.2.0

# AI and Machine Learning
openai==1.3.8
openai-whisper==20230918

# Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and Debugging
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: For better logging and monitoring
structlog==23.2.0
rich==13.7.0